Move Vue components into their own files
which also include their template right there if possible (i.e. if no server-side functions are called)
